Handle exit codes in build script

+++ build.ps1
@@ -33,7 +33,9 @@ $msbuildArgs = @(
 )
 
 & $msbuild /t:build /restore @msbuildArgs
+if ($LastExitCode) { exit 1 }
 & $msbuild /t:pack /p:NoBuild=true @msbuildArgs
+if ($LastExitCode) { exit 1 }
 
 # Test
 dotnet tool install altcover.global --tool-path tools
@@ -53,11 +55,14 @@ foreach ($testAssembly in Get-ChildItem -Recurse -Path src\*.Tests\bin\$configur
     $savedDirectory = '__Saved'
     Remove-Item -Recurse -Force $savedDirectory -ErrorAction Ignore
     & $altcover --inputDirectory=$directory --inplace --outputDirectory=$savedDirectory --opencover --xmlReport=$testResultsDir\coverage.$tfm.xml --assemblyExcludeFilter='AmbientTasks.Tests|NUnit3.TestAdapter'
+    if ($LastExitCode) { exit 1 }
     Remove-Item -Recurse -Force $savedDirectory -ErrorAction Ignore
 
     & $vstest $testAssembly /Logger:'console;verbosity=minimal' /Logger:"trx;LogFileName=$tfm.trx" /ResultsDirectory:$testResultsDir
+    if ($LastExitCode) { $testsFailed = true }
 
     & $altcover runner --collect --recorderDirectory=$directory
+    if ($LastExitCode) { exit 1 }
 
     if ($env:CODECOV_TOKEN) {
         # Workaround for https://github.com/codecov/codecov-exe/issues/71
@@ -65,8 +70,11 @@ foreach ($testAssembly in Get-ChildItem -Recurse -Path src\*.Tests\bin\$configur
         Push-Location $testResultsDir
 
         & $codecovFullPath --name $tfm --file coverage.$tfm.xml --token $env:CODECOV_TOKEN
+        if ($LastExitCode) { exit 1 }
 
         # Workaround for https://github.com/codecov/codecov-exe/issues/71
         Pop-Location
     }
 }
+
+if ($testsFailed) { exit 1 }
